patch TracerProvider
